+++ src/main/java/com/mango/amango/domain/auth/controller/AuthController.java
@@ -3,7 +3,10 @@ package com.mango.amango.domain.auth.controller;
 import com.mango.amango.domain.auth.entity.dto.request.LoginReq;
 import com.mango.amango.domain.auth.entity.dto.request.SignUpReq;
 import com.mango.amango.domain.auth.entity.dto.response.TokenRes;
-import com.mango.amango.domain.auth.service.AuthService;
+import com.mango.amango.domain.auth.service.LogoutService;
+import com.mango.amango.domain.auth.service.ReissueTokenService;
+import com.mango.amango.domain.auth.service.SignInService;
+import com.mango.amango.domain.auth.service.SignUpService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
@@ -16,29 +19,32 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/auth")
 public class AuthController {
 
-    private final AuthService authService;
+    private final SignInService signInService;
+    private final SignUpService signUpService;
+    private final LogoutService logoutService;
+    private final ReissueTokenService reissueTokenService;
 
     @PostMapping
     public ResponseEntity<Void> signUp(@Valid @RequestBody SignUpReq request) {
-        authService.signUp(request);
+        signUpService.signUp(request);
         return ResponseEntity.ok().build();
     }
 
     @PostMapping("/login")
     public ResponseEntity<TokenRes> signIn(@Valid @RequestBody LoginReq request, HttpServletResponse response) {
-        TokenRes tokenRes = authService.signIn(request, response);
+        TokenRes tokenRes = signInService.signIn(request, response);
         return ResponseEntity.ok(tokenRes);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> logout(HttpServletRequest request) {
-        authService.logout(request);
+        logoutService.logout(request);
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping
     public ResponseEntity<TokenRes> reissueToken(HttpServletRequest request) {
-        TokenRes tokenRes = authService.reissueToken(request);
+        TokenRes tokenRes = reissueTokenService.reissueToken(request);
         return ResponseEntity.ok(tokenRes);
     }
 }
