+++ src/main/java/com/mango/amango/domain/auth/service/LogoutService.java
@@ -1,0 +1,8 @@
+package com.mango.amango.domain.auth.service;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface LogoutService {
+
+    void logout(HttpServletRequest request);
+}
