+++ src/main/java/com/mango/amango/domain/auth/service/SignUpService.java
@@ -1,0 +1,8 @@
+package com.mango.amango.domain.auth.service;
+
+import com.mango.amango.domain.auth.entity.dto.request.SignUpReq;
+
+public interface SignUpService {
+
+    void signUp(SignUpReq request);
+}
