+++ src/main/java/com/mango/amango/domain/auth/service/impl/ReissueTokenServiceImpl.java
@@ -1,9 +1,7 @@
 package com.mango.amango.domain.auth.service.impl;
 
-import com.mango.amango.domain.auth.entity.dto.request.LoginReq;
-import com.mango.amango.domain.auth.entity.dto.request.SignUpReq;
 import com.mango.amango.domain.auth.entity.dto.response.TokenRes;
-import com.mango.amango.domain.auth.service.AuthService;
+import com.mango.amango.domain.auth.service.ReissueTokenService;
 import com.mango.amango.domain.user.entity.User;
 import com.mango.amango.domain.user.service.UserService;
 import com.mango.amango.global.exception.CustomErrorCode;
@@ -11,50 +9,17 @@ import com.mango.amango.global.exception.CustomException;
 import com.mango.amango.global.security.jwt.JwtRule;
 import com.mango.amango.global.security.jwt.service.JwtService;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class JwtAuthServiceImpl implements AuthService {
+public class ReissueTokenServiceImpl implements ReissueTokenService {
 
     private final UserService userService;
     private final JwtService jwtService;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public TokenRes signIn(LoginReq request, HttpServletResponse response) {
-        User reqeustUser = userService.findUserByEmail(request.getEmail());
-
-        if (!passwordEncoder.matches(request.getPassword(), reqeustUser.getPassword())) {
-            throw new CustomException(CustomErrorCode.INVALID_PASSWORD);
-        }
-
-        return jwtService.generateTokenResponse(reqeustUser);
-    }
-
-    @Override
-    public void signUp(SignUpReq request) {
-
-        //TODO 이메일 인증
-
-        userService.saveUser(User.builder()
-                .email(request.getEmail())
-                .nickname(request.getNickName())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .build()
-        );
-    }
-
-    @Override
-    public void logout(HttpServletRequest request) {
-        User currentUser = userService.getCurrentUser();
-        jwtService.logout(currentUser);
-    }
 
     @Override
     public TokenRes reissueToken(HttpServletRequest request) {
